+++ lib/dawn/knowledge_base.rb
@@ -267,6 +267,7 @@ require "dawn/kb/cve_2016_0752"
 require "dawn/kb/cve_2016_0753"
 require "dawn/kb/cve_2016_2097"
 require "dawn/kb/cve_2016_2098"
+require "dawn/kb/cve_2016_2785"
 require "dawn/kb/cve_2016_3693"
 require "dawn/kb/cve_2016_5697"
 require "dawn/kb/cve_2016_6316"
@@ -575,6 +576,7 @@ module Dawn
           Dawn::Kb::CVE_2016_0753.new,
           Dawn::Kb::CVE_2016_2097.new,
           Dawn::Kb::CVE_2016_2098.new,
+          Dawn::Kb::CVE_2016_2785.new,
           Dawn::Kb::CVE_2016_3693.new,
           Dawn::Kb::CVE_2016_5697.new,
           Dawn::Kb::CVE_2016_6316.new,